refactor: moved anyValue function to helpers file
