+++ core/plugins/stack/helpers.ts
@@ -1,0 +1,3 @@
+export function anyValue(records: Record<string, unknown>): boolean {
+  return Object.values(records).some((v) => v);
+}
