+++ core/plugins/stack/javascript/linters.ts
@@ -1,4 +1,5 @@
-import { anyValue, IntrospectFn } from "../../../types.ts";
+import { IntrospectFn } from "../../../types.ts";
+import { anyValue } from "../helpers.ts";
 import {
   ESLint,
   introspect as introspectESLint,
